requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.30.6
gunicorn==21.2.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.5.2
python-dotenv==1.0.1
email-validator==2.3.0
alembic==1.13.2
pytest==8.3.3
httpx==0.27.2
